Size!!

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int w;
 	float y;
 
-	printf("Size of a char:%lu byte(s)\n", sizeof(a));
-	printf("Size of an int:%lu bytes(s)\n", sizeof(a));
-	printf("Size of a long int: %lu bytes(s)\n", sizeof(k));
-	printf("Size of a long long int: %lu bytes(s)\n", sizeof(w));
-	printf("Size of a float: %lu bytes(s)\n", sizeof(y));
+	printf("Size of a char:%lu byte(s)\n", (unsigned long) sizeof(a));
+	printf("Size of an int:%lu bytes(s)\n", (unsigned long) sizeof(a));
+	printf("Size of a long int: %lu bytes(s)\n", (unsigned long) sizeof(k));
+	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long) sizeof(w));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned long) sizeof(y));
 	return (0);
 }
